requirements/production.txt
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.2.0
Pillow==10.0.1
requests==2.31.0
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
django-maintenance-mode==0.15.0
